requirements.txt
future~=0.18.2
Pillow~=10.4
requests~=2.25.1
urllib3~=1.26.4
pycryptodome~=3.20.0